Moving Util::Stubs under the Util::Assertion namespace
This commit moves the existing stubs class underneath a new module in
the Util namespace intended to contain all spec related util classes.
